requirements.txt
numpy
Pillow
tensorflow
keras
streamlit